+++ packages/yivite-cli/vite.config.js
@@ -8,7 +8,7 @@ import Components from 'unplugin-vue-components/vite';
 import * as ComponentResolvers from 'unplugin-vue-components/resolvers';
 import { visualizer as Visualizer } from 'rollup-plugin-visualizer';
 import vueI18n from '@intlify/unplugin-vue-i18n/vite';
-import { chunkSplitPlugin as ChunkSplit } from 'vite-plugin-chunk-split';
+import { chunkSplitPlugin as ChunkSplit } from '@yicode/yivite-chunk';
 // import ImageMin from 'unplugin-imagemin/vite';
 import DefineOptions from 'unplugin-vue-define-options/vite';
 // import { viteZip as ZipFile } from 'vite-plugin-zip-file';